Rename and restructure crate and binaries
