+++ src/server/main.rs
@@ -1,0 +1,8 @@
+extern crate velox;
+extern crate rand;
+
+mod serv;
+
+fn main() {
+    serv::Server::new().run()
+}
